+++ Payload_Type/atlas/agent_code/Atlas.csproj
@@ -83,11 +83,19 @@
   <ItemGroup>
     <Compile Include="Config.cs" />
     <Compile Include="Crypto.cs" />
+    <Compile Include="Dispatch.cs" />
     <Compile Include="Http.cs" />
-    <Compile Include="Modules.cs" />
+    <Compile Include="Messages.cs" />
+    <Compile Include="Modules\AssemblyManagement.cs" />
+    <Compile Include="Modules\ConfigManagement.cs" />
+    <Compile Include="Modules\FileManagement.cs" />
+    <Compile Include="Modules\JobManagement.cs" />
+    <Compile Include="Modules\PatchBuffer.cs" />
+    <Compile Include="Modules\ProcessManagement.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils.cs" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
