+++ Payload_Type/atlas/agent_code/Atlas.sln
@@ -8,7 +8,8 @@ EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
@@ -16,12 +17,16 @@ Global
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Debug|x64.ActiveCfg = Debug|x64
+		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Debug|x64.Build.0 = Debug|x64
+		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Debug|x86.ActiveCfg = Debug|x86
+		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Debug|x86.Build.0 = Debug|x86
 		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|Any CPU.Build.0 = Release|Any CPU
-        {91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|x86.ActiveCfg = Release|x86
-		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|x86.Build.0 = Release|x86
-        {91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|x64.ActiveCfg = Release|x64
+		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|x64.ActiveCfg = Release|x64
 		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|x64.Build.0 = Release|x64
+		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|x86.ActiveCfg = Release|x86
+		{91F7A9DA-F045-4239-A1E9-487FFDD65986}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
